Gather search functions in the same file
