+++ src/components/search/SearchResults.js
@@ -1,17 +1,9 @@
 import React from 'react'
 import { withRouter } from 'react-router-dom'
-// import PropTypes from 'prop-types'
 import { useSelector } from 'react-redux'
-import orderBy from 'lodash.orderby'
 
-import searchForProjects from '../../selectors/search'
-import {
-  allProjects,
-  sortBy,
-  getFullProject,
-  getTagsById
-} from '../../selectors'
-import { PaginationProvider, paginateItemList } from '../common/pagination'
+import { allProjects, getTagsById } from '../../selectors'
+import { PaginationProvider } from '../common/pagination'
 import Spinner from '../common/Spinner'
 import { Button, PageTitle, EmptyContent } from '../core'
 import MainContent from '../common/MainContent'
@@ -21,6 +13,7 @@ import { useSearch } from './SearchProvider'
 import { updateLocation } from './search-utils'
 import { TagIcon } from '../core/icons'
 import { getProjectSelectorByKey } from '../../selectors/project'
+import { findProjects } from './find-projects'
 
 export const SearchResultsContainer = () => {
   const { selectedTags, query, sortOption, page } = useSearch()
@@ -107,43 +100,6 @@ const SearchResultsTitle = ({ query, selectedTags, total }) => {
   )
 }
 
-function findProjects(
-  projects,
-  tagsById,
-  auth,
-  { tags, query, page = 1, selector, limit }
-) {
-  console.info('Find', tags, query, page)
-  const filterByTag = project =>
-    tags.every(tagId => project.tags.includes(tagId))
-
-  const filteredProjects = projects.filter(project => {
-    if (tags.length > 0) {
-      if (!filterByTag(project)) return false
-    }
-    return true
-  })
-
-  const foundProjects = query
-    ? searchForProjects(filteredProjects, query)
-    : filteredProjects
-
-  const sortedProjects = false ? foundProjects : sortBy(foundProjects, selector)
-
-  const relevantTags =
-    (tags.length > 0 || query) && getResultRelevantTags(sortedProjects, tags)
-
-  const paginatedProjects = paginateItemList(sortedProjects, page, { limit })
-
-  const results = paginatedProjects.map(getFullProject(tagsById, auth))
-
-  return {
-    results,
-    relevantTags,
-    total: foundProjects.length
-  }
-}
-
 const showCount = (total, text) => {
   if (total === 0) return `no ${text}`
   return `${total} ${text}${total > 1 ? 's' : ''}`
@@ -207,31 +163,6 @@ const NoProjectsFound = withRouter(
     )
   }
 )
-
-function getResultRelevantTags(projects, excludedTags = []) {
-  const projectCountByTag = getTagsFromProjects(projects, excludedTags)
-  return orderBy(
-    Array.from(projectCountByTag.entries()),
-    ([tagId, count]) => count,
-    'desc'
-  )
-}
-
-function getTagsFromProjects(projects, excludedTagIds = []) {
-  const result = new Map()
-  projects.forEach(project => {
-    project.tags
-      .filter(tagId => !excludedTagIds.includes(tagId))
-      .forEach(tagId => {
-        if (result.has(tagId)) {
-          result.set(tagId, result.get(tagId) + 1)
-        } else {
-          result.set(tagId, 1)
-        }
-      })
-  })
-  return result
-}
 
 const RelevantTags = ({ tagIds, baseTagIds }) => {
   const tags = useSelector(getTagsById(tagIds))
